Write user details to db from login component
